requirements.txt
asgiref==3.9.1
Django==5.2.4
django-smart-selects==1.7.2
sqlparse==0.5.3
tzdata==2025.2
